requirements.txt
maritalk==0.2.6
python-dotenv>=1.0.1
tensorflow>=2.0
newsapi-python
datetime